Move `Matcher`s in `refaster-support` from `.util` to `.matchers`
