+++ refaster-support/src/main/java/tech/picnic/errorprone/refaster/matchers/IsArray.java
@@ -1,4 +1,4 @@
-package tech.picnic.errorprone.refaster.util;
+package tech.picnic.errorprone.refaster.matchers;
 
 import static com.google.errorprone.matchers.Matchers.isArrayType;
