+++ refaster-support/src/main/java/tech/picnic/errorprone/refaster/matchers/package-info.java
@@ -6,4 +6,4 @@
  */
 @com.google.errorprone.annotations.CheckReturnValue
 @javax.annotation.ParametersAreNonnullByDefault
-package tech.picnic.errorprone.refaster.util;
+package tech.picnic.errorprone.refaster.matchers;
