+++ refaster-support/src/test/java/tech/picnic/errorprone/refaster/matchers/IsArrayTest.java
@@ -1,4 +1,4 @@
-package tech.picnic.errorprone.refaster.util;
+package tech.picnic.errorprone.refaster.matchers;
 
 import static com.google.errorprone.BugPattern.SeverityLevel.ERROR;
 
@@ -7,49 +7,49 @@ import com.google.errorprone.CompilationTestHelper;
 import com.google.errorprone.bugpatterns.BugChecker;
 import org.junit.jupiter.api.Test;
 
-final class IsCharacterTest {
+final class IsArrayTest {
   @Test
   void matches() {
     CompilationTestHelper.newInstance(MatcherTestChecker.class, getClass())
         .addSourceLines(
             "A.java",
             "class A {",
-            "  String negative1() {",
-            "    return \"a\";",
+            "  Object negative1() {",
+            "    return alwaysNull();",
             "  }",
             "",
-            "  char[] negative2() {",
-            "    return \"a\".toCharArray();",
+            "  String negative2() {",
+            "    return alwaysNull();",
             "  }",
             "",
-            "  byte negative3() {",
-            "    return (byte) 0;",
+            "  int negative3() {",
+            "    return alwaysNull();",
             "  }",
             "",
-            "  int negative4() {",
-            "    return 0;",
-            "  }",
-            "",
-            "  char positive1() {",
+            "  Object[] positive1() {",
             "    // BUG: Diagnostic contains:",
-            "    return 'a';",
+            "    return alwaysNull();",
             "  }",
             "",
-            "  Character positive2() {",
+            "  String[] positive2() {",
             "    // BUG: Diagnostic contains:",
-            "    return (Character) null;",
+            "    return alwaysNull();",
             "  }",
             "",
-            "  char positive3() {",
+            "  int[] positive3() {",
             "    // BUG: Diagnostic contains:",
-            "    return (char) 1;",
+            "    return alwaysNull();",
+            "  }",
+            "",
+            "  private static <T> T alwaysNull() {",
+            "    return null;",
             "  }",
             "}")
         .doTest();
   }
 
-  /** A {@link BugChecker} that simply delegates to {@link IsCharacter}. */
-  @BugPattern(summary = "Flags expressions matched by `IsCharacter`", severity = ERROR)
+  /** A {@link BugChecker} that simply delegates to {@link IsArray}. */
+  @BugPattern(summary = "Flags expressions matched by `IsArray`", severity = ERROR)
   public static final class MatcherTestChecker extends AbstractMatcherTestChecker {
     private static final long serialVersionUID = 1L;
 
@@ -57,7 +57,7 @@ final class IsCharacterTest {
     // https://github.com/checkstyle/checkstyle/issues/10161#issuecomment-1242732120.
     @SuppressWarnings("RedundantModifier")
     public MatcherTestChecker() {
-      super(new IsCharacter());
+      super(new IsArray());
     }
   }
 }
