+++ refaster-support/src/test/java/tech/picnic/errorprone/refaster/matchers/ThrowsCheckedExceptionTest.java
@@ -1,4 +1,4 @@
-package tech.picnic.errorprone.refaster.util;
+package tech.picnic.errorprone.refaster.matchers;
 
 import static com.google.errorprone.BugPattern.SeverityLevel.ERROR;
